Add crate for reusable x509 chain operations
