+++ Cargo.toml
@@ -34,6 +34,7 @@ anyhow = "1.0"
 base64 = "0.21"
 byteorder = "1.5"
 chrono = { version = "0.4", features = ["serde"] }
+der = "0.7.9"
 ed25519-dalek = "2.1.1"
 rand = "0.8.5"
 serde = { version = "1.0", features = ["derive"] }
