+++ crates/ct_worker/Cargo.toml
@@ -68,6 +68,7 @@ thiserror.workspace = true
 tlog_tiles = "0.2.0"
 tokio = { version = "1", features = ["sync"] }
 worker = { version = "0.5.0" }
+x509_util = { path = "../x509_util" }
 x509-verify.workspace = true
 
 [lints.rust]
