+++ crates/x509_util/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "x509_util"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+license.workspace = true
+readme.workspace = true
+homepage.workspace = true
+repository.workspace = true
+description.workspace = true
+
+[dependencies]
+der.workspace = true
+x509-verify.workspace = true
